+++ py-polars/polars/internals/type_aliases.py
